A strong/exact computer of 2^2^k
Generator of Petri nets which are strong/exact computers of double exponent 2^2^k. Constructs of R.J.Lipton & J.Esparza have been minimized and modified. A block contains 23 places and 19 transition. Key "-p" adds 4 priority arcs to each block making the net an exact computer. In other words it allows to guess a successful branch of a nondeterministic computation and transform it into a deterministic (exact) computation. For k=4, when running in stepper simulator of Tina, it looks like a "busy beaver" for Petri nets.